requirements.txt
asgiref==3.5.0
Django==3.2.12
Pillow==9.0.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
Unipath==1.1
